requirements.txt
# Flask i podstawowe zależności
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0

# Apify
apify-client>=1.0.0

# Gemini
google-generativeai>=0.3.0

# Przetwarzanie danych
pandas>=2.0.0

# Wizualizacje
matplotlib>=3.7.0

# Raporty
markdown>=3.4.0
xhtml2pdf>=0.2.13
python-docx>=1.0.0

# Narzędzia
requests>=2.31.0
python-dateutil>=2.8.0

